+++ CHANGELOG.md
@@ -1,21 +1,33 @@
 # Changelog
 
+## [0.2.0] - 2025-09-29
+
+### Added
+
+- Live UI reload when config settings change
+
+### Changed
+
+- Explicitly set the id of the `Harmony` instance to the plugin GUID
+
+- Improve wording of _some_ documentation notes
+
 ## [0.1.4] - 2025-09-27
 
 ### Changed
 
-- Remove mention of how to manually the mod on `Steam Linux`
+- Remove mention of how to install `BepInEx` on `Steam Linux`
 
 BepInEx `v5.4.23.4` does not longer need a custom `doorstop` script, so
-the installation guide provided in the website sufficient (for the most
-part...)
+the installation guide provided in the website is sufficient (for the
+most part...)
 
 - Add comments about how `ShowAllIfAnyCompleted` and `ShowAll` display
 works
 
 ### Fixed
 
-- Changelog links to version `1.0.2` and `1.0.3`
+- Changelog links to version `0.1.2` and `0.1.3`
 
 ## [0.1.3] - 2025-09-26
 
@@ -41,6 +53,7 @@ works
 
 - Initial functionality
 
+[0.2.0]: https://github.com/UserIsntAvailable/Silksong.EndingIndicators/compare/v0.1.4...v0.2.0
 [0.1.4]: https://github.com/UserIsntAvailable/Silksong.EndingIndicators/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/UserIsntAvailable/Silksong.EndingIndicators/compare/1eded41de90d40527d2a405e0b5cc26ba35ef555...v0.1.3
 [0.1.2]: https://github.com/UserIsntAvailable/Silksong.EndingIndicators/compare/v0.1.1...1eded41de90d40527d2a405e0b5cc26ba35ef555